refact(InfoEmployee.jsx, helpers.js, navigation.js): passando funções de formatação para helpers.js, renomeando navigation.jsx para navigation.js
