+++ src/pages/checkout/Movement/CheckoutTable.jsx
@@ -8,12 +8,7 @@ import { useForm } from "react-hook-form";
 import { useParams } from "react-router-dom";
 import movementCheckout from "./MovementCheckout.json";
 import MovementType from "./MovementType";
-import {
-  Button,
-  MonetaryInput,
-  DialogBox,
-  LayoutMenu,
-} from "../../../components";
+import { Button, MonetaryInput, DialogBox } from "../../../components";
 import "./style.scss";
 import { useNavigate } from "react-router-dom";
 import { formatCurrency } from "../../../utils/helpers";
