+++ src/pages/checkout/Movement/MovementType.jsx
@@ -1,10 +1,10 @@
-const MovementType={
-    PIX: "pix",
-    DINHEIRO: "dinheiro",
-    CARTAO_DEBITO: "cartão debito",
-    CARTAO_CREDITO: "cartão credito",
-    CHEQUE: "cheque",
-    GYMPASS: "gympass"
+const MovementType = {
+  PIX: "pix",
+  DINHEIRO: "dinheiro",
+  CARTAO_DEBITO: "cartão debito",
+  CARTAO_CREDITO: "cartão credito",
+  CHEQUE: "cheque",
+  GYMPASS: "gympass",
 };
 
 export default MovementType;
