+++ src/utils/navigation.js
@@ -9,3 +9,7 @@ export const goView = (navigate, routeName, id) => {
 export const goEdit = (navigate, routeName, id) => {
   navigate(`/${routeName}/${id}/editar`);
 };
+
+export const goOrigin = (navigate, routeName) => {
+  navigate(`/${routeName}`);
+};
